requirements.txt
python-docx
googletrans==3.1.0a0
nltk
tqdm